Split memory-related block functions into mem.h
